JSON Placeholder Project
The goal of the project was to implement the association between posts and comments. I displayed the data with the help of ant design components. I also added the requested feature of being able to add a comment to the any posts and save it to the state. I added tests to functions that are essential to the use of the components and any new components I made.